+++ loft-taxi/src/index.js
@@ -6,41 +6,22 @@ import * as serviceWorker from './serviceWorker';
 import {applyMiddleware, compose, createStore} from "redux";
 import {Provider} from "react-redux";
 import rootReducer from "./store/reducers";
-import {authMiddleware, registrMiddleware, logOutMiddleware} from "./store/middlewares";
 import {setLogOut, setUserAcitve, setUserToken} from "./store/actions";
 import {BrowserRouter as Router} from "react-router-dom";
-
 import createSagaMiddleWare from 'redux-saga';
-import {
-    authorizationSaga,
-    registrationSaga,
-    logOutSaga,
-    addressListSaga,
-    paymentSaga,
-    getPayInfoSaga,
-    routeSaga
-} from './store/sagas'
+import rootSaga from "./store/rootSaga";
 
 const sagaMiddleware = createSagaMiddleWare();
 
-
 const store = createStore(
     rootReducer,
-    // compose(applyMiddleware(authMiddleware, registrMiddleware, logOutMiddleware),
     compose(applyMiddleware(sagaMiddleware),
         window.__REDUX_DEVTOOLS_EXTENSION__
             ? window.__REDUX_DEVTOOLS_EXTENSION__()
             : noop => noop
     ));
 
-sagaMiddleware.run(authorizationSaga);
-sagaMiddleware.run(registrationSaga);
-sagaMiddleware.run(logOutSaga);
-sagaMiddleware.run(addressListSaga);
-sagaMiddleware.run(paymentSaga);
-sagaMiddleware.run(getPayInfoSaga);
-sagaMiddleware.run(routeSaga);
-
+sagaMiddleware.run(rootSaga);
 
 const stringInfo = JSON.parse(localStorage.getItem('userInfo'));
