+++ loft-taxi/src/scss/_header.scss
@@ -6,14 +6,6 @@
   z-index: 100;
   margin: 0 auto;
 
-  &__user {
-    background: $ell;
-    padding: 3px 10px;
-    border-radius: 10px;
-    font-size: 13px;
-    margin-right: -120px;
-  }
-
   &__wrap {
     width: 1200px;
     display: flex;
@@ -22,6 +14,15 @@
     height: 100%;
     margin: 0 auto;
   }
+
+  &__user {
+    background: $ell;
+    padding: 3px 10px;
+    border-radius: 10px;
+    font-size: 13px;
+    margin-right: -120px;
+  }
+
 
   &__right {
     display: flex;
@@ -65,5 +66,20 @@
     height: 38px;
     cursor: pointer;
     outline: none;
+  }
+}
+
+
+@media (max-width: 1240px) {
+  .header {
+    &__wrap {
+      box-sizing: border-box;
+      width: calc(100% - 32px);
+      display: flex;
+      justify-content: space-between;
+      align-items: center;
+      height: 100%;
+      margin: 0 $sp;
+    }
   }
 }
